[Op] Align op namespace with DataAPI. Separating op attr headers
