+++ python/tvm/relax/op/__init__.py
@@ -20,11 +20,11 @@
 # Operators
 from .base import *
 from .create import *
+from .datatype import *
 from .manipulate import *
 from .op_attrs import *
-from .reduce import *
+from .statistical import *
 from .tensor import *
-from .transform import *
 from . import builtin
 from . import image
 from . import memory
