+++ src/relax/op/tensor/manipulate.cc
@@ -22,6 +22,8 @@
  * \brief Manipulation operators.
  */
 
+#include <tvm/relax/attrs/manipulate.h>
+
 #include <numeric>
 
 #include "../op_common.h"
